refactor: isolate Cucumber and Postman scenarios processing
