+++ server/src/main/java/com/decathlon/ara/test/ScenariosIndexer.java
@@ -1,0 +1,12 @@
+package com.decathlon.ara.test;
+
+import com.decathlon.ara.domain.ExecutedScenario;
+import com.decathlon.ara.domain.Run;
+
+import java.io.File;
+import java.util.List;
+
+public interface ScenariosIndexer {
+
+    List<ExecutedScenario> getExecutedScenarios(File parentFolder, Run run);
+}
